changed DI to be more DRY
