+++ src/Accountant/Accountant.csproj
@@ -23,9 +23,10 @@
     <PackageReference Include="EntityFramework" Version="6.4.4"/>
     <PackageReference Include="Microsoft.Extensions.Hosting" Version="7.0.*"/>
     <PackageReference Include="Microsoft.AspNet.WebApi.Client" Version="6.0.*"/>
-    <PackageReference Include="Microsoft.AspNetCore.OpenApi" Version="7.0.*"/>
-    <PackageReference Include="GilGoblin.Database" Version="0.6.*"/>
-    <PackageReference Include="GilGoblin.Api" Version="0.4.0-alpha"/>
+<!--    <PackageReference Include="GilGoblin.Api" Version="0.5.*"/>-->
+  </ItemGroup>
+  <ItemGroup>
+    <ProjectReference Include="../Api/Api.csproj"/>
   </ItemGroup>
   <ItemGroup>
     <None Update="appsettings.json" CopyToPublishDirectory="PreserveNewest" CopyToOutputDirectory="PreserveNewest"/>
