+++ src/Api/Api.csproj
@@ -20,11 +20,11 @@
     <PackageReference Include="Serilog.Sinks.Console" Version="4.1.*"/>
     <PackageReference Include="Serilog.Sinks.File" Version="5.0.*"/>
     <PackageReference Include="Serilog.Sinks.RollingFile" Version="3.3.*"/>
-    <PackageReference Include="EntityFramework" Version="6.4.4" />
-    <PackageReference Include="Microsoft.AspNet.WebApi.Client" Version="6.0.0" />
+    <PackageReference Include="EntityFramework" Version="6.4.4"/>
+    <PackageReference Include="Microsoft.AspNet.WebApi.Client" Version="6.0.*"/>
     <PackageReference Include="Microsoft.AspNetCore.OpenApi" Version="6.0.*"/>
-    <PackageReference Include="Swashbuckle.AspNetCore" Version="6.4.0"/>
-    <PackageReference Include="GilGoblin.Database" Version="0.6.*" />
+    <PackageReference Include="Swashbuckle.AspNetCore" Version="6.4.*"/>
+    <!--    <PackageReference Include="GilGoblin.Database" Version="0.6.*"/>-->
   </ItemGroup>
   <ItemGroup>
     <None Update="appsettings.json" CopyToOutputDirectory="PreserveNewest"/>
@@ -32,5 +32,7 @@
     <None Include="LICENSE.txt" Pack="true" PackagePath="/"/>
     <None Include="icon.png" Pack="true" PackagePath="/"/>
   </ItemGroup>
-
+  <ItemGroup>
+    <ProjectReference Include="../Database/Database.csproj"/>
+  </ItemGroup>
 </Project>
