+++ src/Fetcher/Fetcher.csproj
@@ -20,7 +20,10 @@
     <PackageReference Include="Serilog.Sinks.File" Version="5.0.*"/>
     <PackageReference Include="Serilog.Sinks.RollingFile" Version="3.3.*"/>
     <PackageReference Include="EntityFramework" Version="6.4.4" />
-    <PackageReference Include="GilGoblin.Database" Version="0.6.*"/>
+<!--    <PackageReference Include="GilGoblin.Database" Version="0.6.*"/>-->
+  </ItemGroup>
+  <ItemGroup>
+    <ProjectReference Include="../Database/Database.csproj"/>
   </ItemGroup>
   <ItemGroup>
     <None Include="README.md" Pack="true" PackagePath="/"/>
