+++ tests/Accountant/DependencyInjectionTests.cs
@@ -2,7 +2,9 @@ using System;
 using GilGoblin.Database;
 using GilGoblin.Database.Pocos;
 using GilGoblin.Accountant;
+using GilGoblin.Api.Cache;
 using GilGoblin.Api.Crafting;
+using GilGoblin.Api.Repository;
 using Microsoft.AspNetCore.Mvc.Testing;
 using Microsoft.Extensions.DependencyInjection;
 using NUnit.Framework;
@@ -27,10 +29,22 @@ public class AccountantDependencyInjectionTests
         Assert.That(client, Is.Not.Null);
     }
 
-    [TestCase(typeof(IDataSaver<RecipeCostPoco>))]
-    [TestCase(typeof(IDataSaver<RecipeProfitPoco>))]
     [TestCase(typeof(IAccountant<RecipeCostPoco>))]
+    [TestCase(typeof(IItemCache))]
+    [TestCase(typeof(IPriceCache))]
+    [TestCase(typeof(IRecipeCache))]
+    [TestCase(typeof(IItemRecipeCache))]
+    [TestCase(typeof(ICraftCache))]
+    [TestCase(typeof(IRecipeCostCache))]
+    [TestCase(typeof(IRecipeProfitCache))]
     [TestCase(typeof(ICraftingCalculator))]
+    [TestCase(typeof(ICraftRepository<CraftSummaryPoco>))]
+    [TestCase(typeof(IRecipeGrocer))]
+    [TestCase(typeof(IPriceRepository<PricePoco>))]
+    [TestCase(typeof(IItemRepository))]
+    [TestCase(typeof(IRecipeRepository))]
+    [TestCase(typeof(IRecipeCostRepository))]
+    [TestCase(typeof(IRecipeProfitRepository))]
     public void GivenAGoblinService_WhenWeSetup_ThenTheServiceIsResolved(Type serviceType)
     {
         using var scope = _factory.Services.CreateScope();
